feat: Dropped coc :smirk:
COC was getting a bit slow in big projects, Angular work was the main
problem. I tried ALE, but it felt like a COC v2, even though it was
faster it felt like changing for the sake of changing.

Vim-lsp still has rough edges. Even when using vim-lsp-settings, the
easier configuration doesn't really work with Deno/Node projects.

As far as the autocomplete, it works most of the times. Auto import is
not really a feature.