feat: Better mergetool/git diff, cleaner .vimrc
Moved Plugins to near the end of the `.vimrc` to load some nice defaults
of vim before.

Moved backups/undodir/swpdir of vim for the system tmp. It makes sure I
don't leak important details (/tmp is auto deleted on reboot) and avoids
error in case your $HOME is encrypted.