feat: New functions and improved existing ones
Added a `cd()` function to auto list the directory in which we are `cd`'ing
into.

Added `__docker_defaults()` function that will be used by all our docker
wrapper commands as default arguments. It auto checks for `.env` file in
the current directory to pass to the container, exposes a random port to
the container and uses the host network for easier dev work.